test(apis/web/api_test): fix unittests
