--- src/dashboard/apigateway/apigateway/tests/apps/api_test/test_utils.py
@@ -1,66 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# TencentBlueKing is pleased to support the open source community by making
-# 蓝鲸智云 - API 网关(BlueKing - APIGateway) available.
-# Copyright (C) 2017 THL A29 Limited, a Tencent company. All rights reserved.
-# Licensed under the MIT License (the "License"); you may not use this file except
-# in compliance with the License. You may obtain a copy of the License at
-#
-#     http://opensource.org/licenses/MIT
-#
-# Unless required by applicable law or agreed to in writing, software distributed under
-# the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
-# either express or implied. See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# We undertake not to change the open source license (MIT license) applicable
-# to the current version of the project delivered to anyone in the future.
-#
-import pytest
-
-from apigateway.apps.api_test.utils import render_path
-
-
-@pytest.mark.parametrize(
-    "params, expected",
-    [
-        (
-            {
-                "path": "/echo/",
-                "path_params": {},
-            },
-            "/echo/",
-        ),
-        (
-            {
-                "path": "/echo/{api_id}",
-                "path_params": {"api_id": "123"},
-            },
-            "/echo/123",
-        ),
-        (
-            {
-                "path": "/echo/{api-id}",
-                "path_params": {"api-id": "123"},
-            },
-            "/echo/123",
-        ),
-        (
-            {
-                "path": "/echo/{api-id}",
-                "path_params": {},
-            },
-            "/echo/{api-id}",
-        ),
-        (
-            {
-                "path": "/echo/",
-                "path_params": {"api_id": 123},
-            },
-            "/echo/",
-        ),
-    ],
-)
-def test_render_path(params, expected):
-    result = render_path(**params)
-    assert result == expected
